requirements.txt
ffmpeg-python==0.2.0
PySimpleGUI==4.60.3
